Add user-permissions tests for NHS number allocation
